pyproject.toml
[project]
name = "wiki"
version = "0.2"
description = "A wiki built on HIVE"
dependencies = [
    "flask",
    "flask_session",
    "beem",
    "redis",
    "bleach",
    "psycopg2-binary"
]

[build-system]
requires = ["flit_core<4"]
build-backend = "flit_core.buildapi"
